CBMC: Add proof for forward NTT
This commit adds a spec and proof for the MLKEM forward NTT.

The main difficulty in proving the forward NTT type-safe is to
formalize the bounds reasoning showing that the additions and
subtractions in the butterflies do not overflow.

Luckily, very basic layer-to-layer bounds reasoning is sufficient
to achieve an output bound of 8*q, which is enough for
reason about the calling code: When moving from layer N to layer N+1,
every coefficient is replaced by the addition or subtractio or a
layer N coefficient with the result of a Barrett multiplication with
a signed canonical twiddle factor. Since such Barrett multiplication
always has absolute value < q (this could be improved significantly
if we needed to), the absolute bounds increase by at most q.

Spelling out the invariants in CBMC is cumbersome and rather verbose.
To facilitate, and to keep runtimes down, this commit splits the NTT
original NTT code into three functions, one per NTT loop:
- The top-level poly_ntt() iterates over the layers, calling ntt_layer()
- ntt_layer computes one NTT layer, calling ntt_butterfly_block() for
  each block of butterflies.
- ntt_butterfly_block() actually does the modular arithmetic.

While this makes the loop-and-feel of the NTT rather different
from what it was, structurally it is almost the same as in the
reference implementation.

Previously, the NTT code included a runtime check for the tighter
bound 5*q for the output of the NTT. The bound and the check are
removed in this commit: While they are still true, it is confusing
to runtime check for a different value than what we are proving.
Instead, a comment is left.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>