CBMC: Clean up clang-format off
Clang-format messes with our CBMC annotation, so we need to turn it off
using clang-format off and clang-format on comments.
We have been a bit inconsistent where to place those comments.
I think it is best to place the clang-format off comment at
the end of the last line that should be formatted (e.g., the function
signature or the loop.
I changed all places to that now.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>