CBMC: Remove index variable from ARRAY_IN_BOUNDS
The helper macro ARRAY_IN_BOUNDS unfolds to a CBMC annotation
asserting that the coefficients of an array are within specified
bounds.

Previously, this macro would require the caller to explicitly state
the type and name of the index variable, even though this would not
actually feature anywhere else in the macro. The original reason for
this was that CBMC does not like the same variable name be used twice
in block of annotations (precondition, postcondition, loop invariants).

This commit simplifies ARRAY_IN_BOUNDS by always picking `int` as the
index type, and auto-generating the variable name. Name clashes are
avoided by internally appending the line number to the index variable.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>