Fix EC2 failures when Github runner name is already taken
I suspect that our CI failures may be due to the runner
registration failing due to the name already being used.
By default, it is using the hostname as the Github runner
name which defaults to the private IP address on EC2
instances. When instances get terminated, it does not
delete the runners in Github which means we now have
a lot of offline runners listed.

This may be explaining why our CI sometimes timesouts
due to the runner not being registered.

I patched ec2-github-runner to add the `--replace` flag
when registering the runner.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>