CBMC: Add spec + proof for poly_invntt_tomont
This commit adds a memory and type safety proof for
the inverse NTT, using CBMC.

Contrary to the proof for the forward NTT, which required
introducing helper functions per loop, the inverse NTT can
be verified as-is, providing suitable invariants for the
inner loops, and unrolling the outer loop. The decisive
simplification compared to the forward NTT is that there's
a constant bound on the coefficients, independent of the
layer. This is because the inverse NTT reduces more eagerly
than the forward NTT.

A stumbling stone during development was the meaning of
the UNWINDSET. In the context of the inverse NTT, the
loop structure is

```c
for (...) { ...}

for (...) { // XXX
    for (...) {
        for (...) { ...}
    }
}
```

and, oddly enough, the loop `XXX` has index 3.

That aside, CBMC+Z3 get through the proof pretty quickly
(<3min on an Apple M1).

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>