avx2: Fix buffer over-read in rej_uniform native code
The AVX2 implementation for rej_uniform operates in chunks of
48 bytes a time. Initially, those are read into two AVX2 registers
of 32 bytes each. The second of those reads goes past the usable
data in the input buffer by 8 byte, which has to be accommodated by
the caller by providing a suitably padded buffer, thereby making the
contract more complicated and non-intuitive.

This commit instead modifies the AVX2 implementation of rej_uniform
to not read past the end of the usable data in the buffer: It replaces
the read at offset 24 (which reads up to byte 56, 8 bytes past the
guaranteed length of 48 bytes) by a read at offset 16, which stays
within the bounds of the buffer. Luckily, even in the previous code,
there is already a permutation right after the load, so adjusting for
the different offset in the load merely means adjusting the indices
in the permutation. In particular, the change does not come at a
performance cost.

A similar issue exists in the second part of the AVX2 implementation
of rej_uniform, which operates in steps of 12 bytes, but reads 16
bytes of the data buffer. This is resolved by iterating only up to
24 bytes before the end of the input buffer. The drop in performance
is negligible.

This fix also allows to remove the padding previously introduced
in `gen_matrix_entry_x4`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>