CI: Make EC2 instance-bringup job non-cancellable
The `start-ec2-runner` job's responsibility is to bring up an EC2
instance and have it self-register as a self-hosted github runner.
The output of this job is needed in the cleanup job `stop-ec2-runner`.

The `stop-ec2-runner` job is marked as `always()`, to make it non-cancellable.
However, it is still possible that `start-ec2-runner` is cancelled after
an instance has been brought up and potentially self-registered as a
GH runner, and in that case, `stop-ec2-runner` wouldn't have the instance
and runner ID at its disposal to clean things up.

This commit tries to fix this race condition by marking `start-ec2-runner`
as `always()`, thereby making it non-cancellable.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>