CI: Benchmark x86_64-based EC2 instances using PMU instead of PERF
Previous benchmarks on x86_64-based EC2 instances indicated some
accuracy in the performance measurements using the `perf` kernel
module. Performance measurements using `rdtsc`, in turn, seem to
work more reliably.

This commit changes the benchmarking CI for x86_64 instances
to use `rdtsc` to obtain the cycle counter, and no longer `perf`.

To allow this the choice between PMU and PERF first needs to be
made an additional parameter to the EC2 benchmarking workflow.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>