ci: Add benchmark regression alert
The threshold is set to 103%, which means that if the previous benchmark
result is 100 cycles, then there should be an GH comment alert in the PR if
the benchmark result is > 103 cycles.

For the `comment-on-alert` to work on PR, we cannot skip the `store
benchmark result` step.
We could avoid pushing to `gh-pages` branch directly from PR by configuring
`auto-push` to `false` if not on `main` branch.

Signed-off-by: Thing-han, Lim <15379156+potsrevennil@users.noreply.github.com>