Pull poly_{,de}compress_du into function from polyvec_{,de}compress
In ML-KEM, we need 3 types of compressions and their resp.
decompression: Compress_du, Compress_dv, and Compress_1.

Currently Compress_dv is implemented in poly_compress,
Compress_du is implemented in polyvec_compress (as it's only
needed for vectors), and Compress_1 is implemented in
poly_tomsg.
This is somewhat confusing.

This commit is the frist step to unify this by splitting
up polyvec_compress and polyvec_decompress and having them
call new functions called poly_compress_du and poly_decompress_du.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>