CBMC: Remove all loop annotations from polyvec - unroll instead
We can remove all loop annotations in the relatively simple
loops in polyvec.c.
I very much prefer fewer annotations in code.

The reason, this didn't actually work before was that we got
the loop index wrong. Functions like POLYVEC_UBOUND are
defined as an empty while loop in debug.h and, hence, count
as a loop for CBMC.

There are potentially more places where this can help.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>