CBMC: Use SMT-lib arrays for slow polyvec_add proof
In what appears to be proof brittleness, the previous commit
changing headers in the polyvec_add file CBMC proof harness
seems to have led to a major increase in proof runtime of
the polyvec_add function.

This commit fixes this by changing the CBMC flags for this
function to not model arrays directly via SMT-lib arrays.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>