Prepare moving verify.c functions to backend
Currently, mlkem-native has the `verify`, `cmov`, and `cmov_int16`
functions in the verify.c function similar as the reference implementation.
By putting them in a separate compilation unit, we stop the compiler
from turning this code into variable time code.

While being portable, this has various downside:
- If LTO were to be used, this can lead to timing side-channels.
  We stop that by not using LTO for verify.c in our own build, but
  we have no control over how other people compile our code -
  this feels dangerous.
- If these functions are in a separte compilation unit, the compiler
  cannot inline them. This is very expensive particularly for `cmov_int16`.
  By writing that gadget in inline assembly, we can allow the compiler to
  inline these calls, but still guarantee that they execute in constant-time.

This commit prepares the switch to a native implementation of the
verify.c functions in a architecture-specific header file.
For now these are not implemented in inline assembly as I wanted to
gather feedback on this first.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>