AArch64: Add secure native implementations for verify.h
This commit adds native implementations of the verify.h API for AArch64.

Instead of writing the functions in assembly, they are largely
kept in C, and only the decisive operations are implemented in
inline assembly to prevent the compiler from inadvertently
introducing secret-dependent branches.

Sticking to C and inline assembly keeps most of the work to the
compiler while giving it large scheduling freedom through the
precise register usage annotations in the inline assembly.
A function call, in turn, would need to obey the function call ABI
and potentially force the compiler to stash registers on the stack.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>