tests: make run_as_root and exec_wrapper a shared options for all tests commands
If the cross prefix differs from the native toolchain and a custom exec
wrapper is provided, the test script may attempt to run the binary using
the custom exec wrapper in QEMU, which is likely unintended.

It'd probably be cleaner and simpler that QEMU be set as the default exec wrapper
when a cross prefix is specified, and no custom exec wrapper is provided.

Signed-off-by: Thing-han, Lim <15379156+potsrevennil@users.noreply.github.com>