Harden constant time functions against inlining and optimizations
The functions in `verify.c` are suspectible to be compiled into variable-time
code when inlined. For this reason, we disable LTO for `verify.c` in our own
build, and explicitly warn about it in the README.md and `verify.c`.

Yet, with mlkem-native being intended as a code-package, there is risk that
users will copy-paste the code and compile it with LTO enabled, overlooking
the warning.

This commit is an attempt to improve robustness towards LTO by rewriting
verify.c so that it is safe even when inlined.

The main idea is to mask values which the compiler can detect to be 0 or 1,
to avoid branches being introduced for them. More precisely, we introduce
a "value barrier" which is a functionally a no-op, but is written so that
the compiler cannot reason that it is. Then, all 0/1 values are fed through
a value barrier before further processing.

We consider two approaches to implement a value barrier:
- When supported, an empty inline asm block which marks the target value
  as clobbered.
- XOR'ing with the value of a volatile global that's set to 0.

The first approach is rather cheap because it only prevents the compiler
from reasoning about the value of the variable past the barrier. The second
generates redundant loads which therefore incur a runtime cost. Yet, it
seems both more portable and more robust.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>