C90: Don't use inline in C90 builds
C90 does not have the inline compiler directive yet.
This commits moves it to a macro so that, we don't use it in C90 builds.
However, in that case the compiler warns about some inline functions in header files not being
used in every compilation unit that includes that header.
To work around it we silence that warning in that case using __attribute__((unused)).

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>