AVX2: Document bounds in NTT and invNTT
This commit documents the AVX2 assembly for the [inv]NTT in the x86_64
native backend. In particular, it tracks the bounds of data through
the various layers.

The invNTT implementation is highly aggressive in terms of minimizing
the number of modular reductions, which makes a pen-and-paper analysis
rather difficult.

At the last invNTT layer, the bounds reasoning applied is not enough to
show absence of overflow. To remedy, an additional reduction is added.

The analysis for the forward NTT is straightforward.

Ultimately, we confirm the absolute bound of 8*MLKEM_Q for the output
of forward and inverse NTT; this is the contractual bound that the
higher-level C-code is working with.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>