C90: debug.c: set _ISOC99_SOURCE before cinlduing stdio.h
When compiling with -std=c90, stdio.h does not declare
snprintf. We only use that in debug.c which I'd claim we
don't need to have C90-compliant as it doesn't have to be
used by our consumers.
By setting _ISOC99_SOURCE, we can still nudge
stdio.h to declare snprintf.
The alternative would be to use sprintf.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>