AVX2: Simplify non-overflow reasoning in inverse NTT
The AVX2 inverse NTT aggressively minimizes the number of modular
reductions required. This is correct to our knowledge, but it is
difficult to reason about: As the previous commit shows, one has
to take into account rather tight bounds for the absolute value
of a Montgomery multiplication.

This commit adds another modular reduction to the AVX2 inverse NTT.
While not strictly necessary, it simplifies the bounds reasoning
considerably, and does not come at a meaningful performance cost.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>