Don't call scalar (de)compression routines by contract
This commit changes the CBMC Makefile's for functions calling
scalar (de)compression routines, to no longer attempt to call
scalar_[de]compress_xxx() by contract. This call-by-contract
mechanism seems to silently not work for `static` and `static inline`
functions, and not marking the functions as `static` seems wrong
in a header file and would only work so long as CBMC deals with
a single compilation unit.

For now, try to just inline the functions in higher level proofs,
and remember to fix this once call-by-contract supports static
functions.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>