Use shake128x4_ctx and shake256x4_ctx instead of keccakx4_state
This commit changes fips202x4.h so that it only exposes the structure
shake128x4_ctx. The previous structure keccakx4_state is removed, and
the Keccak implementation changed to work with uin64_t * instead.

The shake256x4_ctx is the same as shake128x4_ctx, but only needed
internally since we do not expose multi-step shake256 APIs.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>