CI: Use separate AWS region for metal instance benchmark
We benchmark c7i on a metal instance with 96 vCPUs, a few of which
can get us past the vCPU quota for our CI accounts.

This commit changes the AWS region in which to run the c7i metal
benchmark, to spread the load and reduce the likelihood of exceeding
the vCPU quota.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>