CBMC: Add custom reordering loop into CBMC scope
`gen_matrix()` concludes with an entry-wise permutation of polynomial
coefficients, if required by the native backend.

Previously, the loop over the matrix coefficients was conditional on
a custom permutation being present, which would exclude it from the
scope of CBMC.

With this commit, the loop is unconditional, and a contract used for
the custom permutation. If no custom permutation is needed, it is
defined as a no-op which the compiler will automate away.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>