Test mlkem-native on AArch64_be
This commit modifies the nix flake to install an AArch64 big endian
compiler in the ci-cross shell.

It also modifies the CI to cross-(build+test) mlkem-native on aarch64_be.

This confirms that the C frontend and backend are working on big endian
systems. This is important to check since we do have a few cases of
endian-ness dependent code.

The AArch64 backend is currently _not_ compatible with big endian systems,
and already disabled (this is because we use __AARCH64_EL__ as a guarding
architecture flag -- for big endian, it would be __ARCH64_EB__).

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>