Remove REF-CHANGE annotations
Those annotations were previously meant to completely enumerate
all changes from the reference implementation, but we have not
been diligent adding them, so that now numerous changes are not
annotated.

This commit gives up on the idea of annotating every change from
the reference implementation, and removes REF-CHANGE annotations
accordingly.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>