CBMC: Add separate proof for rej_uniform_scalar
Currently, rej_uniform_scalar and rej_uniform are basically the same
if no native backend is used (as is the case with CBMC proofs). However,
once we include the axiomatized call to native backend in the scope of
CBMC, the proof will be simpler if we hoist out the contract for the
scalar fallback.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>