Fix ordering of includes in arith/fips202 backend
The backend API `mlkem/native/api.h` includes sanity
checks on the arithmetic backend. For example, if a custom
NTT domain coefficient ordering is defined, then the backend
must provide native replacements for almost all functions, and
they have to be adapted to the custom order.

Previously, however, `mlkem/native/api.h` was included _beffore_
the backend. This still served the purpose of ensuring that the
native API is indeed what the backend implements, but it rendered
the sanity checks effectless.

This commit fixes this by reordering the includes accordingly,
in both the arithmetic and the FIPS-202 backend.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>