AArch64: Remove ASM_LOAD directives from poly[vec]_{clean/opt}.S
The ASM_LOAD macro loads the address of a symbol into a GPR.
It uses platform-specific assembly, and is thus cumbersome
to work with. It is also unnecessarily complicated if the goal
is a load from a literal pool: In this case, one can use a PC-relative
load directly, rather than using ASM_LOAD to load the address of
the pool in a register first, and then load the actual data.

This commit removes all uses of ASM_LOAD(...) in poly_{clean/opt}.S
and polyvec_{clean/opt}.S by replacing them by PC-relative loads.

Minor cleanups and improvements are made along the way:
- Store duplicated constants to allow to setup a constant vector
  with a single load, rather than going through a GPR + dup.
- Align .req and .unreq directives, and ensure they match

When trying to re-optimize polyvec* using SLOTHY, it was noted
that ld2/st2 do not appear to be supported in the upstream A55
model yet -- perhaps the code was once optimized using a local
patch. For the time being, this commit therefore replaces st2/ld2
by wrappers which do a manual zip/unzip followed by a normal ldr/str.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>