Python 3.7: Switch from yml to json to not rely on pyyaml
The recent pyyaml 6.0.2 dropped support for anything
before Python 3.8. As we want to support Python 3.7,
we cannot use the most recent version of pyyaml.
Using an old version (6.0.1) would be a workaround.

This commit eliminates the dependency on pyyaml.
We only use yaml in a single place to parse META.yml
which contains meta-information (we mostly care
about various testvector hashes that we compare to).
This commit switches that to a json representation
as json parsing is well supported natively in all
Python versions we care about (and we use json
parsing in the ACVP tests anyway).

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>