Python 3.7: Do not rely on enum value evaluating to __str__(..)
For some reason
f"{enum_value}" evaluates to the numeric representation
(e.g, "1" for TEST_TYPES.MLKEM) in Python3.7,
while it evaluates to the string representation (__str__)
in Python >=3.8.

Workaround would to be to add an explicit str(...), but
this commit instead adds an explicit function called
make_target mapping a test type to a make target.
This would also allow us to use a differnt name for
the make target than the name of the enum value.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>